// frontend/src/app/api/generated/model/ward-bed-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WardBedDto { 
    uid?: string;
    no?: string;
    status?: string;
    active?: boolean;
    wardUid?: string;
}