// frontend/src/app/api/generated/model/ward-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WardDto { 
    uid?: string;
    code?: string;
    name?: string;
    noOfBeds?: number;
    active?: boolean;
    wardCategoryUid?: string;
    wardTypeUid?: string;
}